Module Zarr_sync.Storage

module Deferred : Zarr.Types.Deferred with type 'a t = 'a

module MemoryStore : Zarr.Memory.S with module Deferred = Deferred

A blocking I/O in-memory storage backend for Zarr v3 hierarchy.

module ZipStore : Zarr.Zip.S with module Deferred = Deferred

A blocking I/O Zip file storage backend for a Zarr v3 hierarchy.

module FilesystemStore : sig ... end

A blocking I/O local filesystem storage backend for a Zarr v3 hierarchy.