Module Storage.FilesystemStore

A blocking I/O local filesystem storage backend for a Zarr v3 hierarchy.

include Zarr.Storage.S with type 'a io := 'a

type t

The storage type.

module Group : sig ... end

module Array : sig ... end

val hierarchy : t -> Zarr.Node.Array.t list * Zarr.Node.Group.t list

hierarchy t returns p where p is a pair of lists representing all nodes in store t. The first element of the pair is a list of all array nodes, and the second element is a list of all group nodes. This operation returns a pair of empty lists if store t is empty.

raises Parse_error
if any node has invalid node_type metadata.

val clear : t -> unit

clear t clears the store t by deleting all nodes. If the store is already empty, this is a no-op.

val create : ?perm:int -> string -> t

create ~perm dir returns a new filesystem store.

raises Failure
if dir is a directory that already exists.

val open_store : ?perm:int -> string -> t

open_store ~perm dir returns an existing filesystem Zarr store.

raises Failure
if dir is not a Zarr store path.